Compute the mMSE gap value

This function computes the mMSE gap value.

compute.mMSEgap(
  beta,
  Xmodel = "gaussian",
  Ydist = "gaussian",
  sigma_X.list = NULL,
  X = NULL,
  nulls.list = NULL
)

Arguments

beta
a p-dimensional vector, true regression coefficients.

Xmodel
model of the covaraites (default: "gaussian").

Ydist
the type of conditional model.

sigma_X.list
a list of length p, with each element being the variance of the conditional distribution.

X
a n by p matrix, containing all the covariates.

nulls.list
a list of length p, whose element is a (|i2|*K)-dimensional vector, which contains K set of null samples.

Value

A vector of p, with each element being the value of I for a given covariate.

Details

TBD

References

TBD