Core procedure of floodgate

This function produces floodgate LCBs for given fitted mu.

fg.inference(
  S,
  mu_X,
  mu_Xk,
  Y,
  V_mean = NULL,
  useMC = TRUE,
  one.sided = TRUE,
  alevel = 0.05,
  test = "z",
  verbose = TRUE
)

Arguments

S
a list of selected variables.

mu_X
a list of kength |S|, whose element is the matrix of mu(X) with dimension n2-by-1.

mu_Xk
a list of kength |S|, whose element is the matrix of mu(Xk) with dimension n2-by-K or n2-by-1.

Y
a n2 by 1 matrix, containing the response variables of the inferene samples.

V_mean
A vector of length |S|, whose element is the expected conditional variance term Var(Xj |X-j).

useMC
whether to use Monte Carlo estimators of the conditional quantities (default: TRUE).=

one.sided
whether to obtain LCB or p-values via the one-sided way (default: TRUE).

alevel
confidence level (defaul: 0.05).

test
type of the hypothesis test (defaul: "z").

verbose
whether to show intermediate progress (default: FALSE).

Value

A list of three objects. inf.out: a matrix of |S|-by-4, containing the p-values, LCI, UCI and the floodgate LCB for variable in S; S: a list of selected variables.

Details

TBD

See also